GUI: move user presets to its own file
